schema: Refactor validators into JSONLoaders.
The original plan was to wrap up the gojsonschema library into easier to
use domain-specific validators. It turns out that by hiding the
complexity, working with validation errors becomes a huge pain.

Schema libraries have been refactored to just provide
a `gojsonschema.JSONloader` that can be used to validate against
a specific schema.

Signed-off-by: Atlas Kerr <atlaskerr@gmail.com>